+++ rds-20140815/alibabacloud_rds20140815/client.py
@@ -3526,6 +3526,8 @@ class Client(OpenApiClient):
             query['VPCId'] = request.vpcid
         if not UtilClient.is_unset(request.v_switch_id):
             query['VSwitchId'] = request.v_switch_id
+        if not UtilClient.is_unset(request.whitelist_template_list):
+            query['WhitelistTemplateList'] = request.whitelist_template_list
         if not UtilClient.is_unset(request.zone_id):
             query['ZoneId'] = request.zone_id
         if not UtilClient.is_unset(request.zone_id_slave_1):
@@ -3687,6 +3689,8 @@ class Client(OpenApiClient):
             query['VPCId'] = request.vpcid
         if not UtilClient.is_unset(request.v_switch_id):
             query['VSwitchId'] = request.v_switch_id
+        if not UtilClient.is_unset(request.whitelist_template_list):
+            query['WhitelistTemplateList'] = request.whitelist_template_list
         if not UtilClient.is_unset(request.zone_id):
             query['ZoneId'] = request.zone_id
         if not UtilClient.is_unset(request.zone_id_slave_1):
@@ -32630,6 +32634,8 @@ class Client(OpenApiClient):
             query['ResourceOwnerAccount'] = request.resource_owner_account
         if not UtilClient.is_unset(request.resource_owner_id):
             query['ResourceOwnerId'] = request.resource_owner_id
+        if not UtilClient.is_unset(request.v_switch_ids):
+            query['VSwitchIds'] = request.v_switch_ids
         req = open_api_models.OpenApiRequest(
             query=OpenApiUtilClient.query(query)
         )
@@ -32687,6 +32693,8 @@ class Client(OpenApiClient):
             query['ResourceOwnerAccount'] = request.resource_owner_account
         if not UtilClient.is_unset(request.resource_owner_id):
             query['ResourceOwnerId'] = request.resource_owner_id
+        if not UtilClient.is_unset(request.v_switch_ids):
+            query['VSwitchIds'] = request.v_switch_ids
         req = open_api_models.OpenApiRequest(
             query=OpenApiUtilClient.query(query)
         )